+++ src/main/java/com/xact/assessment/repositories/OrganisationRepository.java
@@ -5,5 +5,5 @@ import io.micronaut.data.annotation.Repository;
 import io.micronaut.data.repository.CrudRepository;
 
 @Repository
-public interface OrganisationRepository extends CrudRepository<Organisation, Long> {
+public interface OrganisationRepository extends CrudRepository<Organisation, Integer> {
 }
